rebase fix: move test code in the 'unit' subfolder
