+++ test/unit/component/parser/xml/XMLTestUtils.ts
@@ -20,10 +20,10 @@ export function verifyIsNotArray(object: any, message?: string): void {
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function verifyBounds(shape: any, x: number, y: number, width: number, height: number): void {
   const bounds = shape.Bounds;
-  expect(bounds.x).toEqual(x, 'bounds x');
-  expect(bounds.y).toEqual(y, 'bounds y');
-  expect(bounds.width).toEqual(width, 'bounds width');
-  expect(bounds.height).toEqual(height, 'bounds height');
+  expect(bounds.x).toEqual(x);
+  expect(bounds.y).toEqual(y);
+  expect(bounds.width).toEqual(width);
+  expect(bounds.height).toEqual(height);
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -45,6 +45,6 @@ export function verifyAndGetBPMNShape(json: any): any {
   const plane = diagram.BPMNPlane;
   verifyProperties(plane, ['BPMNShape'], []);
   const shapes = plane.BPMNShape;
-  verifyIsNotEmptyArray(shapes, 'BPMNShape');
+  verifyIsNotEmptyArray(shapes);
   return shapes;
 }
